--- langchain_app/CCC_scraper/requirements.txt
@@ -1,5 +1,0 @@
-bs4
-jq
-langchain
-langchain-community
-requests
