--- langchain_app/requirements.txt
@@ -1,14 +1,0 @@
-bs4
-langchain
-langchainhub
-langchain_anthropic
-langchain-chroma
-langchain-community
-langchain_google_genai
-langchain-experimental
-langchain_openai
-langgraph
-pymupdf
-pypdf
-requests
-unidecode
