+++ src/indigobot/CCC_scraper/requirements.txt
@@ -1,0 +1,5 @@
+bs4
+jq
+langchain
+langchain-community
+requests
