+++ src/indigobot/requirements.txt
@@ -1,0 +1,14 @@
+bs4
+langchain
+langchainhub
+langchain_anthropic
+langchain-chroma
+langchain-community
+langchain_google_genai
+langchain-experimental
+langchain_openai
+langgraph
+pymupdf
+pypdf
+requests
+unidecode
