+++ tests/test_sql_agent.py
@@ -6,7 +6,7 @@ from unittest.mock import Mock, patch
 
 from langchain_core.documents import Document
 
-from langchain_app.sql_agent import (
+from indigobot.sql_agent import (
     DB_PATH,
     format_docs,
     load_docs,
